Implemented the level Nik designed
Created the level Nik designed as best I could. Obviously, there are certain entities that have not yet been created, so they aren't in the level, but there's space for them. I also had to change the design a bit, since the spacing in the level wasn't quite plausible with the dimensions of the character. This can be changed to be more accurate to the design later if desired.