Restructure of files, scene transitions
Restructured the level files and code that handles the menus.

Elevators are an item on each level, they are a prefab that can be put anywhere. upon collision with the player they take you to the next level in the build list. Current 1-3 does not take you anywhere (it probably crashes tbh).

Worked with scene transitions too which was a pain, try to avoid the built-in Animator if you can.