Added player jumping sound and polished a few minor things
Added one of the jumping sounds to the player, which plays whenever it jumps. Added a folder which contains the .wav files. I also edited the second and third level so that the player's jumping physics matched the first level. I also moved the elevators in the second and third levels to the actual end so the player couldn't just cheat their way to the end.