Sprint 4
Worked on elevator, its not perfect but I was messing with our moving platforms to simulate us rising with the elevator.

Made a jump pad that interacts with the player and spear, the floor jump (bounce) pads interact with both but the walls only interact with the spear. Could be useful for combat levels later or trying to get the player to shoot around a couple walls for a button.

Tested different spear velocities, I personally like 1400 but left it at 800 like it was.

On level 1-1 I tested a cursor, easily able to be disabled if we do not like it.

Made a rudimentary pushable block, could get it to interact with a button if need be.

Made it so that you can also retrieve your spear if you press "e".