Added Alex's bgm
Alex's music now plays on loop in each of the levels. I tried to balance the audio so it wasn't too loud or quiet, but if it needs to be changed, that's fine. The only issue is that it starts over  when the player dies. The music is tied to the camera, so when the level restarts, so does the music, and I don't know how to get around this.